Make Command+BACKSPACE delete all previous characters
When interacting with the OmniBar, I noticed that pressing
Command+BACKSPACE when at the end of the line only removed the final
character instead of the expected behavior of removing all of the
previous characters. Oddly, it seemed to behave as expected if the
cursor was not at the very end of the line.

To resolve this, I added a special case for Command+BACKSPACE to clear
all previous characters. I don't have enough context yet to determine if
this is the correct way to solve this problem, or even if it is actually
a problem at all, but it seems like a decent improvement to me.