Downgrade Babel
So, async/await is broken with the update I tried to apply in d9b16da. I
tried to bisect this, but Node's recursive dependencies makes it super
difficult.

It looks the generated output is the same, so I am suspicious of one of
the upgraded dependencies (regenerator, perhaps), but when I downgrade
it (by hand-editing `npm-shrinkwrap.json`, blowing away `node_modules`
and re-installing) things still don't work.

When I `npm link` in my babel checkout, then `npm link babel-core` in
corpus, `git bisect` reports that everything is wonderful throughout the
entire commit range (the puzzling part is, it is... stuff works).

So I guess I am a n00b and don't know how to bisect this kind of thing
in a node world, and I am giving up.

Note that this is not a straight revert, because 87c6ecf had changed the
JSON file. This commit is effectively the result of:

```
git co 69b4e73 -- npm-shrinkwrap.json
rm -r node_modules/*
npm install
npm update eslint
```

(ie. remove the eslint and babel updates, then re-apply the eslint
update)