Reset NoteList scrollTop on clearing selection
We were already doing this when hitting ESCAPE, but we need to do it
when the selection goes away for other reasons too. For example, in this
scenario:

- open app, top note is selected
- search for "cheatsheet"; it matches notes but is not an exact title
  match, so we clear the selection

Because we were failing to reset the scrollTop in this case, we
effectively wound up scrolled all the way to the bottom during this
search, which is a poor user experience.