Make bubbling animations work when note list is filtered
Quoting my prophetic self from d6a3488683524c8a239bf9fa:

    Probably a hint that I should refactor to avoid those confusing
    names; here `index` can mean "position with filtered notes list" or
    "position within global notes list".

That commit fixed a bug where we were using "position within the
filtered notes list" to identify notes within the global notes list,
causing the wrong note to be written to.

In doing this, we broke animations for notes in a filtered view, because
we'd animate "from" the position within the global notes list (usually
further down) rather than within the filtered list.

The solution is to propagate both indices in the payload, and use them
accordingly:

- `index` is the index within the global store and is used for writing
- `position` is the position with the visible filtered notes list, and
  is used for animating

Everything works, and the names still suck, and there is probably
another future bug (or two) lurking here waiting to be born.