Fix ESCAPE/BACKSPACE behavior in OmniBar
Another chapter in the story of the teetering tower of instability that
is the state machine of the application, this commit fixes some issues
with the BACKSPACE key in the OmniBar. Hopefully it doesn't introduce
any other bugs in the process.

What we were seeing:

- enter a prefix of an existing note: "Keto [Journal]"
- hit BACKSPACE, the note list would get inappropriately refiltered

This was caused because we were initiating a new search (from the
`NoteList`) whenever the selection cleared. Not good. Backspacing clears
the selection.

So, don't do that. Let the OmniBar worry about searching.

After doing this, the ESCAPE key was a little busted, so I waved a dead
chicken over the codebase until it started working again.