Blur #private note previews rather than omitting them
Playing with this idea. Not sure if it is a good one. Omitting the
preview looks semi-broken and very bare, but this does not. It's
possible that it is distracting though. I will run with it for a while
and see how it feels.

Using `text-shadow` + `transparent` to get the blur, as `-webkit-filter:
blur(...)` is not blurry enough.