Avoid problems caused by OS X open file limit
Found some views blowing up because of accessing non-existent properties
like "text". Digging deeper, I saw that the cause was OS X's low file
limits causing many opens to die with "ENFILE: file table overflow",
which we were swallowing silently.

Hadn't noticed this before, so I think there is something local on my
system that must be using a log of file handles right now. (Even before
the recent changes to the note store, these swallowed errors would have
resulted in empty notes, which should have been visible in the UI).

So, do stuff in smaller batches, avoiding the errors. And when errors do
occur, make sure they don't propagate up to views by filtering out the
dud items. This is better than fleshing out partial dummy items, for the
reasons noted in the comment inline.

Currently `console.error`-ing to communicate that this happened, but I
will create an issue for building out a proper UI for showing that there
are errors (could be an exclamation icon, and clicking on it could open
the console, if I even need an explicit action for it).