Port LogStore to be a reducer
We know have a `logs` reducer, with a facade that presents to the rest
of the app as `LogStore`. Picked this as a first one to port because it
is super easy.

Note I also changed the event names from retrospective to imperative,
which seems to be the blessed way to do it with Redux. When the dust
settles will decide what to do with my "action creators". There is
overlap there between `log.{warn,error}` and the creators in
`Actions.js`. I am not sure where the side-effect of calling
`console.warn` or `console.error` should go, but the docs seem to
suggest that async or side-effectful action creators should be ok.