Connect ContentEditable to Redux-powered config
Wrap it, making the exported value a container. Does this mean I should
move it into the "containers/" directory (or should I reserve that for
containers that have enough of their own logic to warrant living in a
separate file)?

Note: this is the only component that talked to `ConfigStore` directly;
all the other users are in the existing Flux stores.