Apply alternative fix for note creation focus regression
This is an alternative and probably less sketchy fix for the regression
originally corrected in a2c5d406db1fa. I'm trying to port these Flux
stores over to Redux reducers one by one, and eliminating asynchrony (in
this case, the `setImmediate` call) is an important step.

(Alas, I think `waitFor` is a spanner in the works, because it means you
can't convert any reducer that uses it, instead having to traverse up
the dependency parent chain until you hit the root...)