Place cursor at top of note by default
Without this, it appears at the bottom, which can be disorienting for
long notes, where the cursor won't even be visible on the screen.

For a better user experience, we'll want to persist cursor position info
across restarts. I've moved that out of `TODO.md` and into the issue
tracker here:

https://github.com/wincent/corpus/issues/40