Replace SystemStore with reducer-based version
Still feeling may way around how to do this gradually, but what this
does is replace `SystemStore` with a reducer.

`SystemStore` gets decomposed into:

- Pure `system` reducer.
- Asynchronous `querySystem` module.

The reducer itself does almost nothing; it really just provides
defaults. I am not sure if it makes much sense to isolate something so
tiny. Anyway...

Our saga is responsible for running the async `querySystem` bit, and it
does so after we load the config. The config is passed in explicitly
from the saga, encoding the dependency in the call graph instead of in
the `SystemStore` (which had to listen for a `CONFIG_LOADED` action). I
am not sure if the `SYSTEM_INFO_LOADED` action is useful or necessary. I
will probably end up removing it.

Note that the one component that cares about the contents of the store
(`OmniBar`) now gets it via a `system` prop. One nice thing about this
is it removes the subscribe/unsubscribe noise from the component. In
other words, the components are becoming cleaner, while we're
accumulating a lot of boilerplate-y code elsewhere. This may be ok, as
the separation of concerns (especially thanks to the sagas) may lead us
to a more maintainable result overall.

I can see a pattern emerging here, in which a lot of the high-level
components in my app are going to be "containers" in the sense mentioned
in the Redux docs. I am undecided yet whether I should move them all
into the "containers" folder or keep them where they are; I kind of like
how the wrapping is internal to the file (via `connect`).