Port LogStore to Redux
Here I am really stretching the limit. This is the least hideous of all
the hideous options that I could think of for making this happen.

Most uses of `log` in our app are inside components, where we in turn
have easy access to `dispatch`. It is not so easy in lower-level
modules, however. We are left with the unsavory choice of living without
the desired feature, making it exist in parallel outside of Redux using
some other management technology, keeping things as standard as possible
but threading through a bunch of references to the application state
and/or dispatcher, or having a global.

The global seems the lesser of all evils, so that's the one we're going
with for now. If I wake up tomorrow and hate this I am going to blame
sleep deprivation.

Component use cases should be easy to clean up from here. There are some
Flux stores that are currently using `log`, but those will all get
ported to Redux and hopefully go away.