Get gulp to put the files where we want them
Using this one cool trick.

From:

  http://stackoverflow.com/a/32880264/2103996

Specifically, this makes:

  dist/foo

Appear at:

  Corpus.app/Contents/Resources/app/dist/foo

Instead of at:

  Corpus.app/Contents/Resources/app/foo

The non-glob entry (package.json), was already appearing at the right
place.

As noted in that Stack Overflow answer, this approach is not 100%
robust, but it works for my case (assumes I only have one file at that
level with the "dist" prefix, and one with the "vendor" prefix).