Work around other view state bugs
This was broken because we weren't dealing with real DOM nodes. The ref
we get here is actually a (Redux) Connect instance. If I remove that, we
still don't work, because the ref is a `ContentEditable` instance. Only
way to get this to work for now is to use `findDOMNode` to get the real
`textarea`.

I'm going to rework this to expose that in some way, which will allow me
to remove the `findDOMNode` calls.