Don't select note when clicking on a tag
This is a better UX. Otherwise you might end up causing the list to be
filtered, cause the notes to move, and then you end up selecting a note
other than the one that you clicked on.

The noted race in the comment is still intermittently present. Seems
that if you have a tag search active (eg. "#private") and that is
showing in the OmniBar, then clicking on another tag will perform the
search but not update the bar; if you have a specific item title showing
in the bar then clicking on a tag does update the bar.