Remove note about unexpected events
According to ImmutableJS docs, `map` always returns a new list, even if
every callback invocation produces the same value. That explains why we
always get an event and the `===` check after filtering is pointless.

Additionally, get rid of a `toList()` that was added back in efdd645671
and which I believe is unnecessary.