Fix lints
Crazy old eslint. Really don't get why it only groks those exclusions if
they use C-style comments. (Line based exclusions seem to work fine with
C99 comments, but note that I can't force Prettier to keep line-based
comments on the same line as a function declaration, so this is the only
way to do it).