Log error info when a command fails
Suddenly every Git operation started failing. I had been lazy with the
implementation here, not bothering to log anything given that I didn't
actually expect the operations to fail, realistically.

Turns out, Apple has my back there! Because of the `PATH` environment in
which we're running, we end up using `/usr/bin/git` and not the
Homebrew-installed `/usr/local/bin/git`. The latter works fine, as
before, but the former started barfing for every single operation with a
recent update:

    Agreeing to the Xcode/iOS license requires admin privileges,
    please run “sudo xcodebuild -license” and then retry this command.