Warn users when changes made outside of Corpus
Building on the last commit, this one actually warns the user (via the
crudest mechanism possible) when things have changed on the filesystem
by an agent other than Corpus itself.

When you see the "error", click it to see which path changed, and hit
Command-Option-R to reload. If that proves to be reliable, I may
consider auto-reloading (provided there are no local changes).

Closes: https://github.com/wincent/corpus/issues/58