Include reload prompt when out-of-band changes detected
Although that Unicode sure looks ugly in the console...

Incidentally, I looked into trying to determine when auto-reloading is
safe by checking `OperationQueue.size`:

- If an unexpected change occurs and the queue size is > 0, that
  means we have a pending write so it would be a bad idea to reload
  automatically and we should instead prompt the user to do a manual
  reload if they want.
- If an unexpected change occurs and the queue size is 0, we have no
  pending writes, so it would be safe to reload.

In testing this with various race conditions (eg. edit a file in Corpus,
then mutate outside within a second) I found that the logic is basically
unneeded.

- If you change a file in Corpus then externally, you are going to have
  a pending change expected, and when the filesystem notification comes
  in you'll end up considering it legit either way.
- If you change a file externally and then in Corpus, you will have
  received a notification about the external change immediately, in
  which case reloading would have (could have) happened before you made
  your change. And in any case, Corpus writes out changes pretty
  aggressively, so if you received an external notification and chose
  not to reload, any edit you made would get written out to the
  filesystem within 5 seconds anyway.

At least, in practice, I was not able to get Corpus to receive an
unexpected change event whenever the queue size was > 0. I think we
could afford to make reloading the default behavior.