Remove dead clean-up code to reduce clutter
I've always had this in there as a matter of "hygiene" even though I
knew it was never executed. But I am not keen on the clutter. Mitigating
the risk that this will be transplanted into a context where the
assumption that the component never gets unmounted may no longer be
valid.