Handle NOTE_BUBBLED in store.js
Not sure what pattern I want to apply here. These could be called
"actions" or "mutations", but they are just functions that do something
to the store. For now just sticking the mutator function in
Actions.js. Eventually I'll be able to delete all the old actions. Not
sure whether I should be storing these centrally or in individual files,
or locally (where there is only one usage). Want to stop store.js from
getting too big, but can make the decision on how to structure this
latter, once I have a feel for the emerging pattern. Don't want things
to get too complicated, which would undermine the entire purpose of the
move from Flux to Undux.