Add version number to notes
If you rename or otherwise change the content of a note we bump the
version number, which will allow us to do cheap change detection on
notes. This will be useful for filtering purposes: if we do a
before/after comparison (ie. look at previously filtered array and new
array after filtering), we can re-use existing objects which haven't
actually changed, or discard the filtered array if it hasn't actually
changed (eg. I change search query from "foo" to "foobar" and the result
set remains the same).