Partial migration towards recent Undux
This is evidently going to be a big enough reworking that doing it in
small steps is going to hurt; so let's commit the cardinal sin of doing
this in big chunks, each of which corresponds to some amount of
intermediate brokenness. Warts and all: plenty of hacks to make this run
without crashing before rendering anything, as well as unsightly
workarounds and commented-out sections. Currently "working": we get as
far as loading notes and shoving them in the store. Using an rxjs
observable seeing as they arrive in chunks anyway.

Thoughts on this so far:

- Some risk in use API that has just been marked as
  `setFrom_EXPERIMENTAL`.
- Switching away from "deprecated" `createStore` API means that all of
  this becomes totally React-centric now. I am partly ok with that
  because ultimately the React lifecycle *is* the lifecycle of the app
  as perceived by the user so it's ok for the UI to be the "owner" of
  the model layer then. On the other hand, this binding is going to make
  testing model-level logic next to impossible to test, so will evaluate
  this once it's all over.