Wire up logging again
As noted, some song-and-dance required to make sure we can log before
the Undux store is in place. I have support for multiple subscribers in
place, although I don't expect to ever need it. Our sole subscriber is
OmniBar.react.js, which is the one that cares about showing the icon
when new log messages arrive.

In practice, I don't expect log messages to be produced before Undux
starts (because it is tied to component life cycle, and we use that to
kick off the load of everything) but it is still good to accommodate
that possibility.