Add FrozenSet for programming in an immutable style
Not really immutable (only enforced in __DEV__).

Reason I want this: in order to continue ripping out the
NotesSelectionStore I need to be able to rely on object identity in
order to detect changes (long story, will explain later). So rather than
relying on programmer discipline to get it right all the time, make a
FrozenSet class. Flow will complain if we use a normal Set where a
FrozenSet is required, and if we ever attempt mutation, our __DEV__-time
checks will alert us.

This shouldn't be much slower than an ordinary Set because it is not in
a hot code path right now (or perhaps ever); it's just modified when the
selection changes, which is something that runs at human speed and not
red-hot-loop speed.