Add "src/common" directory
The "renderer" and "main" processes are entirely different contexts that
communicate via IPC and share very little code indeed. Make this latter
category (shared code) explicit by hoisting it into a new "top-level"
directory called "common". At the moment it's just `nullthrows()` and
it's possible that that's all it ever will be.

With this structure I will be able to maintain a separate `.flowconfig`
for each of "renderer" and "main" (unfortunately necessary because the
"electron" module presents a different set of exports in each context).