Use separate .flowconfig for main and renderer processes
Unfortunately necessitated because the "electron" module presents
different exports in each context.

Unfortunate because I now have two config files to keep (mostly) in
sync, and because this now means I get two separate sets of `flow` run
output instead of one. Nevertheless, I think it's all worth it in the
name of correctness.

Note the use of "$@" in the wrapper (`scripts/flow`), which means we can
still stuff like:

```
yarn flow --show-all-errors
```