Add back a top-level .flowconfig for the benefit of LSP tools
Sadly, this means that my attempts at creating two separate environments
aren't 100% complete. `yarn flow` will do the right thing from the
command-line, but when I am in Vim I'll get pseudo-correct errors
based on the assumption that all the files are using the "renderer"
version of the "electron" module.