feat: add initial experimentation with previews
Just experimenting with UI patterns at this point. First tried the
approach shown here (custom command mode completion, showing preview
window for selected files) and couldn't get it to work. Then went down
rabbit hole with `getchar()` (also couldn't get it to work).

Discovered that in both case I just need a `:redraw` to force update the
display and show preview, but seeing the number of edge cases I'd need
to handle in the `getchar()` case I backed away a bit.

Issue with the approach as implemented in this commit is it is filename
based instead of content based (obviously), and it is incomplete (also
obviously); but still, I'm pleased that the preview stuff works as well
as it does. Shows promise even if I go back and explore a `getchar()`
type thing again.