feat: add FrozenSet
Took this from the "next" branch. Notably, I couldn't get a subclass to
typecheck in TypeScript, unlike the Flow version, so I made a wrapper
instead that just implements the Set interface. This makes it a bit more
verbose, but all the tests pass without modification.

Note the ugly __DEV__ hack. Asked for suggestions on how to do this
better here:

https://twitter.com/wincent/status/1152551100542803968