fix: remove unneeded type assertions
We can get rid of the `React.CSSProperties` type assertions by using `as
const` instead, which prevents TypeScript from widening our style object
types.

Note the removal of the `JSONObject` assertion which shows why
assertions can be a bad thing: we actually had a bug in here where
`null` or an array would would have passed the `typeof` check, rendering
our assertion invalid; that's a bug, which we fix here and in the
process obviate the need for the assertion.