feat: improve type safety of style declarations
Took me ages to figure this one out. For some reason, the module
augmentation I am doing in css.d.ts didn't work if I did it in
types.d.ts; it needed to be in a separate file (any attempt at doing an
`import` -- any `import` at all -- in types.d.ts breaks `tsc` horribly).
Switching it to a non-ambient file (ie. types.ts instead of types.d.ts)
broke things in other ways.

It also worked if I did it in the two files that need `WebkitAppRegion`,
but I didn't want to duplicate, and it's ugly even in one place. The
alternatives I considered, such as wrapping that up in a `useStyles`
hook, were just too ugly to contemplate. I was just about to add an
`any` cast in disgust, but finally thought of trying a separate file.

Upstream issue requesting this be added:

    https://github.com/frenic/csstype/issues/71

Docs on augmentation:

    https://github.com/frenic/csstype#what-should-i-do-when-i-get-type-errors