feat: tighten up Styles typing
My goal here was to catch typos of this form:

    // styles.button exists, but styles.buton does not:
    <div styles={styles.buton} />

the problem is, the `Styles` type used an index signature:

    type Styles = {
      [name: string]: React.CSSProperties;
    };

which means that even with an `as const` object literal:

    const styles: Styles = {
      button: ...,
    } as const;

the type ends up with that index signature instead of the more precise
desired inferred type:

    type DesiredInferredType = {
      button: React.CSSProperties
    }

    type ActualInferredType = {
      [name: string]: React.CSSProperties
    }

Just as `as const` doesn't help, `Readonly<>` doesn't either. There are
a bunch of ways to get the stricter type, but they all involve
duplication, and if you do it wrong, then you'll have a hole. Some
examples:

     const groups = {
      button: ...,
     };

     const styles: {[k in keyof properties]: ...} = {
      ...groups,
     };

(Sucks because of the extra object allocation and unwanted extra
variable.)

    type Styles<T> = {[k in keyof T]: ...};

    const groups = ...;

    const styles: Styles<typeof groups> = {
      ...groups,
    };

(More "expressive", but still has the unwanted object allocation and
extra variable.)

    function getStyles<T extends {[name: string]: number}>(
      obj: T
    ): {[k in keyof T]: number} {
        return obj;
    }

    const styles = getStyles({
      button: ...,
    });

(No longer has the unwanted spread, but it has a totally useless
function call that does nothing but return its argument, to allow
TypeScript to infer the right type.)

    type Styles = ReturnType<typeof getStyles>;

    const styles: Styles = {
      button: ...,
    };

(Avoids the function call, but now TypeScript doesn't infer the right
type anymore; it goes back to using an index signature.)

    type Style<T extends {[name: string]: number}> = {
      [k in keyof T]: number
    };

    const style: Style<{button: React.CSSProperties}> = {
      button: ...;
    };

(Too verbose.)

Final version ends up using a union of strings instead, which you use
like this:

    const style1: Style<'button'> = {...};
    const style2: Style<'button' | 'root'> = {...};

This feels boilerplatish, but it is about as good as we can do. It
catches all the right errors:

- Forgetting to provide type parameter.
- Providing the wrong number of type parameters.
- Providing the wrong type of type parameter.
- Making a typo of a name in the type parameter.
- Making a typo while accessing a property on the resulting object (ie.
  the actual motivating use case).