refactor: remove sketchy useCallback usage from useStyles hook
Linter is going to complain if we don't include `callback` in the
dependencies list, but `callback` will be different on every render,
which means the memoization is pointless. It is fine to call the
callback every time anyway because it does almost no work. We could
hoist `useCallback()` up to the component if we ever thought it would
get too expensive; eg.

    const styles = useStyles<'root'>(
      useCallback(
        () => ({
          root: {
            prop: focused ? 'a' : 'b',
          },
        }),
        [focused]
    );

but probably never will though. So we have two function calls on every
render (one to `useStyles()` and one to the callback); if perf ever
becomes a real issue we can inline the calculations anyway. May keep
`type Styles` around for that purpose, despite the TODO comment.