refactor: use useReducer for dependent state variables
Based on this Twitter thread, where we discussed the two-booleans vs a
tri-state variable:

https://twitter.com/Randemonium/status/1161365535294132225

Thought I'd give it a try just to evaluate the trade-offs:

- Pros:
  - Dependency between state "atoms" is now centralized (and therefore
    obvious) inside the reducer.
  - Removes a logical OR expression.
  - Removes an `if`/`else if` conditional.
  - Replaces 4 calls to 2 `set` methods with 2 calls to 1 `set` method.
  - Replaces a chained ternary with a single ternary.
- Cons:
  - Adds one complex conditional (4 branches).
  - Adds a logical AND expression.
  - Slightly longer (+3 lines).
  - Requires ugly "RESET" faux-level to signal the need to reset (could
    avoid that by having separate actions -- eg. 'show-error',
    'show-warning' and 'reset' -- but that would make it even longer).

It's still not clear to me which one is better, which probably means
that either is fine. Even listing out the pros and cons like that it is
hard to get a truly objective measure of which one is simpler, or which
one is more likely to have a bug now (or in the future).