fix: make Rollup build correctly deal with ES modules
As explained in that lengthy comment, even though Rollup is made to work
with ES modules, it doesn't interact well with TypeScript.

TS compiles to code which uses the ".js" variants, CommonJS requires,
and explicit access of the `.default` property. At least, that's what it
does with the settings I am using.

So, when Rollup comes along it uses the ES module code directly, but
doesn't get rid of the `default()` access that JS added, so you
effectively wind up with `thing.default.default()` which isn't going to
end well.