chore(corpus): enable Corpus experiment but turn it into a no-op
Because I am still deciding on the approach to take. This is stuff I've
had locally in a plug-in for many months now thinking that I would like
to replace the Corpus Electron app with an in-Vim solution:

https://github.com/wincent/corpus

One of my major gripes that I'd like to tackle with this is that:

- Linking within Markdown kind of sucks; see:
  https://github.com/wincent/masochist/issues/145
- Corpus isn't Vim.

But Corpus has some features that I really like, like easy content and
title searching (with previewing, which is what I was trying to get at
with this experiment).

I think the way forward here might be as simple as setting up some
mappings to make it easy to create articles from links, and update
metadata fields and link footers, and jump to linked articles.

The problem noted in here of title and content searched not jiving well
together can be fixed by adding metadata to each doc and updating titles
on save.