fix: handle edge cases when outside of Corpus directory
Example:

  :Corpus <tab>

Expands to:

  :Corpus /

If you hit enter there, we don't want to blindly `cd` to "/". In fact we
never want to `cd` to anything with a slash, unless it is a valid Corpus
directory.

  :Corpus whatever else

should be taken as a cue to open or create the note "whatever else.md"
in the default corpus directory (the first one in the config).