fix: remember selected file before disposing of buffer
Fixes regression introduced in the last commit.

Overall, a smell here is that we assume non-nil values in a bunch of
places, relying on a fragile implicit "state machine" (ie. chooser
will be in state with non-nil value here because we count on having
previously passed through state that set up the values).