fix: avoid E33 when updating references in new buffer
Say you're in a new buffer and haven't used a `:substitute` command yet.
You save the buffer, and boom:

    E33: No previous substitute regular expression

As `:h E33` states:

    When using the '~' character in a pattern, it is replaced with the
    previously used pattern in a ":substitute" command. This fails when
    no such command has been used yet. See |/~|. This also happens when
    using ":s/pat/%/", where the "%" stands for the previous substitute
    string.

The problem is this call:

    matchlist(l:line, '\%#=1\v^ {0,3}(`{3,}|~{3,})[^`]*$')

Our intent is to match a line starting with 3 backticks or 3 tildes, but
`~` has a special meaning, as noted in `:h /~`:

    ~     matches the last given substitute string

The fix, then, is simple: escape the tilde!

While this is mostly an innocuous bug (producing error noise), it could
potentially do damage too. If there is a previous substitute string, our
match list could potentially contain matches for it, and then the code
that executes after it is effectively running with invalid assumptions.